requirements.txt
flask
numpy==1.15.0
pillow
tensorflow==1.12.0
torch==1.6.0
transformers==3.0.2
seqeval==0.0.12
pytorch-crf==0.7.2
websocket-client==1.3.1
cffi==1.12.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
pycparser==2.19
six==1.16.0